requirements.txt
python-nmap==0.7.1
python3-nmap==1.5.1
simplejson==3.17.6
